+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix no jumping sounds on remote
 - Fix footstep sounds playing when dashing on remote
 - Fix player lock status not working after entering a new map
+- Fix warnings on some combat arts
 
 ## [0.8.3] 2026-08-01